Revert "Separate default feature from has module (#407)"
This reverts commit e9b561d22c14d277f4fd7a441b7b6e6bcea6c753.
