+++ intern.json
@@ -28,7 +28,6 @@
 	],
 
 	"plugins": [
-		"./dist/dev/tests/shim/support/has-plugin.js",
 		{
 			"script": "./dist/dev/src/shim/main",
 			"useLoader": true
@@ -36,7 +35,8 @@
 		{
 			"script": "./dist/dev/tests/core/support/jsdom-plugin.js",
 			"useLoader": true
-		}
+		},
+		"./dist/dev/tests/shim/support/has-plugin.js"
 	],
 
 	"browser": {
