+++ src/shim/main.ts
@@ -1,6 +1,5 @@
 import './tslib';
 
-import '../core/features';
 import * as array from './array';
 import * as iterator from './iterator';
 import Map from './Map';
