+++ tests/has/unit/has.ts
@@ -372,7 +372,6 @@ registerSuite('has', {
 				'es6-string',
 				'es6-string-raw',
 				'es6-symbol',
-				'es6-iterator',
 				'es6-weakmap',
 				'es7-array',
 				'microtasks',
